requirements.txt
flask 
openai 
pandas
IPython
python-dotenv